Time: 491 ms (13.35%), Space: 118 MB (19.63%) - LeetHub

+++ 2472-maximum-number-of-non-overlapping-palindrome-substrings/2472-maximum-number-of-non-overlapping-palindrome-substrings.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maxPalindromes(string s, int k) {
-        using ll = long long;
+        using ll = int;
         ll n = s.length();
         vector<array<ll, 2>> intervals;
         for(ll i = 0; i < n; i++)
